third_party/blink/web_tests/fast/css/counters/after-continuation-expected.txt

This test passes if the numbers displayed below are 1 and 2 in this order.

PASS successfullyParsed is true

TEST COMPLETE

PASS internals.counterValue(document.getElementById('c2')) is '2'


TEST COMPLETE
